requirements.txt
Django==2.0.4
django-debug-toolbar==2.0
django-extensions==2.1.9
django-tenant-schemas==1.9.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
psycopg2==2.8.3
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
